Test hiding with multiple NPCs
Glad the static hiding spot list idea seems to work. Kind of a tidy solution for avoiding managers or searches for this kind of thing. Might want NPCs to run away and hide in the first free spot they see, though, instead of psychically detecting free hiding spots and making a beeline for them lol